Edited db_gen files.
Need to update the genecp part. 

Example .smi and .sdf files attached. 

should be run as python dg_gen.py smi.smi (or) sdf.sdf

As of now will output the com files in a new folder gaussian.

Will have to add all the additional scripts to a new folder. Will do it 
once all the scripts are completed.