Created a separate file for variable and function.
added the output analyser to the db_gen_functions. (so no more need of 
output_analysis.py)

added in arguments for the second run of gaussian calculation

added in lines where the qsub_summit will woork after creating the input 
files.

edited the location of moving files after second run, if normally 
terminated. 

Adding in a bash script with will do the xtb part. following which 
qsub_summit will create the .sh scripts for gaussian.

Will have to test this if it works on summit. 

Will do that later today or tomorrow. 

Work with this versions from now on.