updated the process if analysis and writing the new .com files.
3.It can now do more than two rounds of optimization by automatically 
creating new imput files if all the files are not optimised in two 
rounds.
2. Depending on the number of rounds, if files are nornally optimised, 
it will more to the Fn=inished folder of the first round.

Updated the analysis by maaking sure gencp is only written only if 
needed. 

Updated the resubmit to multiple rounds from only two.

TO DO: Move boltz, combine to second chuck of the Project